change name to inspectBlinks and introduced prepreEnv
The function's purporse is not so much to show the blinks, but to allow the user to inspect the setting of blink related parameters (mainly the "expand-blink" time).

Previous versions used a .GlobalEnv assignment. Created a prepreEnv environment from which ui.R and server.R extract the current ppdt that was set in inspectBlinks().